Fix project master templates
The `project.yml` generated by `nvflare provision` command has
comments `change overseer.example.com to the FQDN of the overseer`
and `change example.com to the FQDN of the server` but the actual
values nowhere has `example.com` mentioned and it is confusing.

This change just fixes comments to make it more clear.