Fix DockerBuilder to add `build` section (#2662)
The `compose.yml` generated by the DockerBuilder adds a build section
only for the overseer. As a result, when we have non-HA mode
`project.yml', generated `compose.yml` won't have the build section and
so `docker compose build` won't buildthe required NVFlare docker image.

This change adds a build section in the generated `compose.yml` even
for clients and servers. It fixes the above issue and also with this
change, same `compose.yml` can be used to build/run individual servers
and clients using the `docker compose` method.