Add error log sending to master template and add provisioning configuration (#3138)
Add error log sending to master template and an ability to configure it
in provisioning.

### Description

Add error log sending to master template and an ability to configure it
in provisioning. A new PropKey is expected for allow_error_sending: true
to be in the project.yml for a client to allow error log sending. If
this is not included, a _modify_error_sender callback added to
StaticFileBuilder's build_from_template will remove the error_log_sender
component from the client during provisioning.

### Types of changes
<!--- Put an `x` in all the boxes that apply, and remove the not
applicable items -->
- [x] Non-breaking change (fix or new feature that would not break
existing functionality).
- [ ] Breaking change (fix or new feature that would cause existing
functionality to change).
- [ ] New tests added to cover the changes.
- [ ] Quick tests passed locally by running `./runtest.sh`.
- [ ] In-line docstrings updated.
- [ ] Documentation updated.