[2.5] Fix conn sec saving (#3170)
Fixes # .

### Description

Currently connection security value is stored to the databus when
creating the client engine. However, CellPipe might be created during
job config file processing (when creating components), which needs the
value of conn security. However, at this time the client engine has not
been created, hence the value is not available.

This PR fixes the problem by saving the conn security value to the
databus in the "start_config" method of the CP's config, which happens
before component creation. This guarantees that the value is available
when CellPipe is created.

### Types of changes
<!--- Put an `x` in all the boxes that apply, and remove the not
applicable items -->
- [x] Non-breaking change (fix or new feature that would not break
existing functionality).
- [ ] Breaking change (fix or new feature that would cause existing
functionality to change).
- [ ] New tests added to cover the changes.
- [ ] Quick tests passed locally by running `./runtest.sh`.
- [ ] In-line docstrings updated.
- [ ] Documentation updated.