Cleanup Bogus Errors (#3190)
Removed several bogus errors:
1. Error processing frame: RuntimeError: cannot schedule new futures
after shutdown
This can happen when the message arrives for a cell already being
shutdown. Changed it to debug.
2. Logical Error: Endpoint is already removed
This is actually not an error. The weak-ref set entry can be removed
when the ref is gone. Changed it to debug.
3. Added name to several threads to help debugging the dangling thread
issue.


### Types of changes
<!--- Put an `x` in all the boxes that apply, and remove the not
applicable items -->
- [x] Non-breaking change (fix or new feature that would not break
existing functionality).
- [ ] Breaking change (fix or new feature that would cause existing
functionality to change).
- [ ] New tests added to cover the changes.
- [ ] Quick tests passed locally by running `./runtest.sh`.
- [ ] In-line docstrings updated.
- [ ] Documentation updated.

---------

Co-authored-by: Ziyue Xu <ziyuex@nvidia.com>