Flower Integration V2 (#3216)
Fixes # .

### Description

This PR implements Flower Integration Version 2 to use Flower's new
architecture.

In Flower's previous architecture, on the server side, there are two
processes: the superlink process and the server-app process. Flare is
responsible for starting and stopping both processes. On client side,
the supernode and client-app are included into a single process, and
Flare is responsible fro starting and stopping this process.

In the new Flower architecture, on the server side, there is a superlink
process, and Flare is responsible for starting and stopping this
process. Once started, Flare issues the "flwr run" command to the
superlink to submit a Flower app. In the meanwhile, Flare issues the
"flwr ls" command periodically to query the status of the submitted app
until the app is finished. On the client side, there is one process
called Supernode, and Flare is responsible for starting and stopping
this process. Flare does not need to start the server-app or client-app
processes - these will be managed by the Flower system.

The overall integration architecture remains the same. The only
difference is how to start and stop Flower apps.

### Types of changes
<!--- Put an `x` in all the boxes that apply, and remove the not
applicable items -->
- [x] Non-breaking change (fix or new feature that would not break
existing functionality).
- [ ] Breaking change (fix or new feature that would cause existing
functionality to change).
- [ ] New tests added to cover the changes.
- [ ] Quick tests passed locally by running `./runtest.sh`.
- [ ] In-line docstrings updated.
- [ ] Documentation updated.