Support custom participant types and custom roles in Provision (#3280)
Fixes # .

### Description

This PR adds back support of custom participant types and custom roles
to the Provision system. These features were in previous versions, but
got removed accidentally during Provision refactoring.

Also added support of client capacity in StaticFileBuilder. This used to
be only available for dashboard in previous versions.

Also improve the handling of race condition that SJ is not ready when CJ
tries to sync.


### Types of changes
<!--- Put an `x` in all the boxes that apply, and remove the not
applicable items -->
- [x] Non-breaking change (fix or new feature that would not break
existing functionality).
- [ ] Breaking change (fix or new feature that would cause existing
functionality to change).
- [ ] New tests added to cover the changes.
- [ ] Quick tests passed locally by running `./runtest.sh`.
- [ ] In-line docstrings updated.
- [ ] Documentation updated.