Enhance CCWF Cross Site Evaluation (#3322)
Fixes # .

### Description

This PR improves the CCWF's cross site evaluation (CSE) to address the
following potential issues:

- The trainer is a 3rd-party program that uses client API and is
launched through a process launcher.
- The model being evaluated is LLM that could take a lot of memory and
communication bandwidth.

Currently, when performing CSE, the Server schedules many parallel tasks
for each evaluator to evaluate each evaluatee. In the case that the
trainer is a 3rd-party program that uses client API and is launched
through a process launcher, these parallel tasks could cause the trainer
to serve multiple model requests from other clients. Since the trainer
uses Client API, which is designed to do one request at a time, it won't
be able to handle those multiple requests. In fact, it only handles one
request and drops all others. This will cause the CSE job to fail.

In case of LLM, multiple model requests at the same time could also
cause communication issues.

This PR implements multiple improvements:
- A new task type "prepare_model" is added. Before scheduling any CSE
tasks, the Server sends the "prepare_model" task to the clients that
have models to be evaluated. Each client will then prepare the model to
be evaluated. For local models, the client will call the trainer program
to get the model and keep it in memory. This way, when clients are
tasked to do cross-site evaluation on these models, the Trainer won't
need to be called again to retrieve their models.
- Instead of scheduling all tasks in one round, the Server uses multiple
rounds to process all required evaluations. In each round, it tries to
schedule as many parallel tasks as possible without causing any site to
be "too busy" (which is controlled by the newly introduced
max_parallel_actions parameter).
- On the client side, when it needs to evaluate its own model, it won't
send a message to itself to retrieve the model; instead it will use the
prepared model directly.


### Types of changes
<!--- Put an `x` in all the boxes that apply, and remove the not
applicable items -->
- [x] Non-breaking change (fix or new feature that would not break
existing functionality).
- [ ] Breaking change (fix or new feature that would cause existing
functionality to change).
- [ ] New tests added to cover the changes.
- [ ] Quick tests passed locally by running `./runtest.sh`.
- [ ] In-line docstrings updated.
- [ ] Documentation updated.

---------

Co-authored-by: Ziyue Xu <ziyuex@nvidia.com>