[2.6] Fix var substitution in config parsing (#3350)
Fixes # .

### Description

Currently we do var substitution for all string type arg values. If the
value happens to contain curly brackets but content within the brackets
is not a valid var name (as reported in
https://github.com/NVIDIA/NVFlare/issues/3229), the substitution will
fail.

This PR adds protection to the substitution: if it fails, the original
value will be returned.

### Types of changes
<!--- Put an `x` in all the boxes that apply, and remove the not
applicable items -->
- [x] Non-breaking change (fix or new feature that would not break
existing functionality).
- [ ] Breaking change (fix or new feature that would cause existing
functionality to change).
- [ ] New tests added to cover the changes.
- [ ] Quick tests passed locally by running `./runtest.sh`.
- [ ] In-line docstrings updated.
- [ ] Documentation updated.