[2.6] Separate STDERR from STDOUT in tie (#3346)
### Issue

While testing the TIE + Flower job, I discovered that if the process
outputs to STDERR, any attempt to use STDOUT for retrieving results
fails when the two streams are mixed together.

For example:
https://github.com/NVIDIA/NVFlare/blob/2.6/nvflare/app_opt/flower/applet.py#L225-L232


### Description

This PR separates STDERR from STDOUT to prevent mixing the two streams
and ensure proper handling of the output.

### Types of changes
<!--- Put an `x` in all the boxes that apply, and remove the not
applicable items -->
- [x] Non-breaking change (fix or new feature that would not break
existing functionality).
- [ ] Breaking change (fix or new feature that would cause existing
functionality to change).
- [ ] New tests added to cover the changes.
- [ ] Quick tests passed locally by running `./runtest.sh`.
- [ ] In-line docstrings updated.
- [ ] Documentation updated.