[2.6] Add custom logger utility (#3365)
Fixes [FLARE-2444](https://jirasw.nvidia.com/browse/FLARE-2444)

For custom user code, we need a way to enforce a naming convention for
loggers to allow these logs through the FLFilter for "concise" log mode.
This is only applicable in the non-clientAPI cases (as for the
clientAPI, print statements are captured and handled)

- Added a `custom_logger` utility function to prepend "custom" to logger
names
- Updated examples and docs accordingly

### Types of changes
<!--- Put an `x` in all the boxes that apply, and remove the not
applicable items -->
- [x] Non-breaking change (fix or new feature that would not break
existing functionality).
- [ ] Breaking change (fix or new feature that would cause existing
functionality to change).
- [ ] New tests added to cover the changes.
- [ ] Quick tests passed locally by running `./runtest.sh`.
- [ ] In-line docstrings updated.
- [ ] Documentation updated.

---------

Co-authored-by: Chester Chen <512707+chesterxgchen@users.noreply.github.com>