Update-based mechanism for edge workflow (#3389)
Fixes # .

### Description

Currently the edge workflow is based on aggregation. This is too
limiting. This PR changes to use a more generic update-based mechanism.
This allows the assessor and updaters to have a wholistic view of the
overall state, which makes it easier for them to manage the workflow.

### Types of changes
<!--- Put an `x` in all the boxes that apply, and remove the not
applicable items -->
- [x] Non-breaking change (fix or new feature that would not break
existing functionality).
- [ ] Breaking change (fix or new feature that would cause existing
functionality to change).
- [ ] New tests added to cover the changes.
- [ ] Quick tests passed locally by running `./runtest.sh`.
- [ ] In-line docstrings updated.
- [ ] Documentation updated.

---------

Co-authored-by: Ziyue Xu <ziyuex@nvidia.com>