Feature/add fedavg metric optrimization controller (#3506)
Fixes # .

### Description

Hi all,

This pull request adds a new controller: **FedAvg Metric Optimizaiton**.
This controller is the result of a joint effort between Roche and
Universitätsspital Zurich (USZ) to train a model in a federated manner.

Purpose of the controller:
The goal is to obtain the best possible model by optimizing a specific
metric (e.g., minimizing the loss or maximizing the F-Score) and to stop
the training if the tracked metric does not improve after a certain
number of FL rounds, as defined by the researcher. This approach saves
computation time during FL training, especially when the model is large
and requires a significant amount of data. This controller has been
developed in a paper that is currently under peer review.

Additionally, we wanted to provide the option to choose whether to
optimize the metric during training or validation.

I would like to highlight that this contribution is thanks to Roche and
the Universitätsspital Zurich (USZ).

Best,
Lydia Anette Schönpflug (USZ) and Ruben Bagan Benavides (Roche)

### Types of changes
<!--- Put an `x` in all the boxes that apply, and remove the not
applicable items -->
- [x] Non-breaking change (fix or new feature that would not break
existing functionality).
- [ ] Breaking change (fix or new feature that would cause existing
functionality to change).
- [ ] New tests added to cover the changes.
- [ ] Quick tests passed locally by running `./runtest.sh`.
- [ ] In-line docstrings updated.
- [ ] Documentation updated.

---------

Co-authored-by: Chester Chen <512707+chesterxgchen@users.noreply.github.com>