Enhance admin functions (#3530)
Fixes # .

### Description

This PR adds following enhancements to the admin functions:
- Use defined constants for admin config properties (instead of
hard-coded strings)
- Make it possible for loading security content (the fed_admin.json) for
multiple users in the same process. This is necessary when the admin API
(flare API or FLAdminAPI) is used for multiple users, and in the
integration tests (which requires multiple user sessions).

### Types of changes
<!--- Put an `x` in all the boxes that apply, and remove the not
applicable items -->
- [x] Non-breaking change (fix or new feature that would not break
existing functionality).
- [ ] Breaking change (fix or new feature that would cause existing
functionality to change).
- [ ] New tests added to cover the changes.
- [ ] Quick tests passed locally by running `./runtest.sh`.
- [ ] In-line docstrings updated.
- [ ] Documentation updated.