Safe tensor decomposer (#3544)
Fixes # .

### Description

This PR address security issue with TensorDecompose. Instead of using
jit.load to serialize BF16 type of tensors (which has security issues),
this PR uses torch.save to save weights.

Note that, in general, torch.save uses Pickle for saving whole model.
BUT we only save model weights (dict of Tensors) and do not save whole
model. This does not involve Pickling.

In addition, this PR makes file-based decomposition a general-purpose
function that can be used to implement decomposers of any large objects.
This is particularly useful when dealing with LLM.

The general ideas are here:

- Objects of target object types (e.g. Tensor, NP Array) contained in
one message are collected into one single dict, and written to a file
using the target object type specific file creation like
savetensors.save_file, or np.save.
- Objects in the payload are replaced with simple references to the
objects written to file. Hence the payload size is hugely reduced.
- The payload is sent to receivers.
- Each receiver downloads file from the sender, using FileDownloader.
- Once downloaded, the payload is recomposed and the object refs are
replaced with received objects in the downloaded file.

With this approach, since file is downloaded with small chunks, the
memory space needed is small (no need to hold and transfer the huge
objects via memory).

This PR adds more enhancement to reduce memory space and increase
efficiency when the same message needs to be sent to multiple receivers
(e.g. in broadcasting task, or message retries in Reliable Message).
Only the first message needs to go thru the file conversion process. All
other messages will use cached result from the 1st message (the
primary).

The core of this solution is the abstract class ViaFileDecomposer.
NumpyArrayDecomposer and TensorDecomposer have been modified to be based
on ViaFileDecomposer.

In order to be able to create file refs during decomposition and
download files during recomposition, cell must be available to
decomposers. To support this, the concept of "FOBS Context" is added,
which is used to contain external information (e.g. cell or other config
params like file download timeout) needed by decomposers.

One issue with this approach is the life cycle management of the
generated temp files - they need to be removed as soon as they are no
longer needed. Two ways are implemented to guarantee that those temp
files are eventually removed.

- One, at the end of the job, always calls FileDownloader.shutdown(),
which closes all pending transactions (which deletes temp files). This
is the last resort that catches all files that somehow didn't get
removed earlier.
- Two, use the concept of "msg root" that lets the app layer manage the
life cycle of the msg. The msg sender first assigns a unique msg root id
to the msg. Then the msg is either sent (to one or multiple receivers),
or waiting for receivers to pull. Eventually, when the msg sender knows
that the msg is no longer needed, it calls delete_msg_root to remove all
temp files associated with the msg root.

In most cases, the msg sender does know when to delete the msg root. The
only exception is "fire and forget" since the msg sender does not wait
for receivers. In this case, Solution One will be able to delete all the
files (if any).

I have updated the following places to use msg root:
- Task-based interactions (wf_comm_server, wf_comm_client,
task_contoller)
- Reliable Message, which may resend the message multiple times
- Task Exchanger, which sends message to client API
- Pipe Handler, which may resend the message multiple times

There might be other places to improve.

### Types of changes
<!--- Put an `x` in all the boxes that apply, and remove the not
applicable items -->
- [x] Non-breaking change (fix or new feature that would not break
existing functionality).
- [ ] Breaking change (fix or new feature that would cause existing
functionality to change).
- [ ] New tests added to cover the changes.
- [ ] Quick tests passed locally by running `./runtest.sh`.
- [ ] In-line docstrings updated.
- [ ] Documentation updated.

---------

Co-authored-by: Chester Chen <512707+chesterxgchen@users.noreply.github.com>