Fix max client limit to exclude admin clients (#3552)
Fixes # .

### Description

This PR fixes the bug 4 (reported in
https://nvbugspro.nvidia.com/bug/5390831) that failed to allow admin
clients once max number of FL clients are connected. Admin clients will
not be counted as FL clients.

This PR also attempts to fix bug 5 in the report. Though we couldn't
reproduce the issue following QA instructions, we found and fixed a
place that could potentially cause this.

This PR also fixes bug 6.


### Types of changes
<!--- Put an `x` in all the boxes that apply, and remove the not
applicable items -->
- [x] Non-breaking change (fix or new feature that would not break
existing functionality).
- [ ] Breaking change (fix or new feature that would cause existing
functionality to change).
- [ ] New tests added to cover the changes.
- [ ] Quick tests passed locally by running `./runtest.sh`.
- [ ] In-line docstrings updated.
- [ ] Documentation updated.