Fix preflight check for clear connection (#3575)
Fixes # .

### Description

This PR fixes the Issue 6 reported in
https://nvbugspro.nvidia.com/bug/5412832.

The preflight check assumes that the connection to the server is always
secure, hence only tries to connect to the server withe secure
connection, which won't work if connection security is set to be clear.

### Types of changes
<!--- Put an `x` in all the boxes that apply, and remove the not
applicable items -->
- [x] Non-breaking change (fix or new feature that would not break
existing functionality).
- [ ] Breaking change (fix or new feature that would cause existing
functionality to change).
- [ ] New tests added to cover the changes.
- [ ] Quick tests passed locally by running `./runtest.sh`.
- [ ] In-line docstrings updated.
- [ ] Documentation updated.

Co-authored-by: Yuan-Ting Hsieh (謝沅廷) <yuantingh@nvidia.com>