Support app packages in job API (#3586)
Fixes # .

### Description

Currently, if a module is not in NVFLARE package, the Job API applies
the magic of determining its class path and copying the source into the
"custom" folder of generated job. When the module is in an app package
that will be included in the final python env (e.g. docker container),
then this magic is not desired. In fact, including the source code could
cause confusion when resolving class path.

This PR implements a simple solution by letting the Job API user specify
app packages to be excluded from this behavior.
If the package name of a module is in the specified app package list,
then the magic won't be applied.


### Types of changes
<!--- Put an `x` in all the boxes that apply, and remove the not
applicable items -->
- [x] Non-breaking change (fix or new feature that would not break
existing functionality).
- [ ] Breaking change (fix or new feature that would cause existing
functionality to change).
- [ ] New tests added to cover the changes.
- [ ] Quick tests passed locally by running `./runtest.sh`.
- [ ] In-line docstrings updated.
- [ ] Documentation updated.