Create BuiltIn FedAvgLR (#3563)
### Description
1. Take the existing LR Newton Raphson method created by @zhijinl , and
put the FedAvgLR as builtin Controller rather than customized code
2. create hello-lr example with client.py, job.py and model.py,
prepare_data.py. The job.py is temporary code with job API, but will be
replaced with JobRecipe code in next phase
3. launch_external_process= True to False to leverage
InProcessClientAPIExecutor. The subprocess launcher is borken for this
example at this point.

A few sentences describing the changes proposed in this pull request.

### Types of changes
<!--- Put an `x` in all the boxes that apply, and remove the not
applicable items -->
- [x] Non-breaking change (fix or new feature that would not break
existing functionality).
- [ ] Breaking change (fix or new feature that would cause existing
functionality to change).
- [ ] New tests added to cover the changes.
- [ ] Quick tests passed locally by running `./runtest.sh`.
- [ ] In-line docstrings updated.
- [ ] Documentation updated.