Add support of msg traffic control for swarm learning (#3604)
Fixes # .

### Description

With Swarm Learning, training clients submit results back to the
aggregation client when they finish training. In most cases, this
happens at roughly the same time. This could cause "traffic jam" on the
aggregation client that gets multiple messages at the same time, which
could cause resource exhaust on the aggregation client, especially when
the training result is huge (e.g. 32GB).

To mitigate this problem, this PR adds "traffic control" to limit the
number of concurrent result submissions. It works as follows:

- Before submitting result, the training client sends "submission
request" to the aggregation client for permission to submit. It does
this repeatedly until a permission is received; or exceeds max wait time
(if specified).
- Once a "submission request" is received, the aggregation client checks
how many concurrent submissions are already in flight. If it's still
below the configured limit, then permission is granted; otherwise
permission is denied and the training client will have to ask again.

Three config args are added to the SwarmClientController:

**request_to_submit_result_msg_timeout** - the timeout for "submission
request" message. Since submission req is a tiny message, this timeout
value should be small. Default to 5.0.

**request_to_submit_result_max_wait** - max amount of time to wait for
the permission from the aggregation client. If the permission is not
received within this period of time, the training result will not be
submitted. If this value is not specified (None), then the training
client will keep trying forever. Default is None.

**max_concurrent_submissions** - max number of concurrent submissions
allowed on the aggregation client. Default is 1.

Note that if the max_concurrent_submissions is set to 1, it practically
means that all training results will be submitted to the aggregation
client sequentially. This lowers the resource pressure on the aggr
client, but makes the overall training process longer. The value of
request_to_submit_result_max_wait, if specified, should be long enough
to allow the aggr client sufficient time to process training results.

### Types of changes
<!--- Put an `x` in all the boxes that apply, and remove the not
applicable items -->
- [x] Non-breaking change (fix or new feature that would not break
existing functionality).
- [ ] Breaking change (fix or new feature that would cause existing
functionality to change).
- [ ] New tests added to cover the changes.
- [ ] Quick tests passed locally by running `./runtest.sh`.
- [ ] In-line docstrings updated.
- [ ] Documentation updated.