Update the token expiration time (#3605)
The token expiration time needs to be less than the check_frequency so
that we ensure the token is fresh (with nonce) everytime we verify it.

### Description

The token expiration time needs to be less than the check_frequency so
that we ensure the token is fresh (with nonce) everytime we verify it.

Note that under the current code, we can only support 1 job at a time
for CC.

### Types of changes
<!--- Put an `x` in all the boxes that apply, and remove the not
applicable items -->
- [x] Non-breaking change (fix or new feature that would not break
existing functionality).
- [ ] Breaking change (fix or new feature that would cause existing
functionality to change).
- [ ] New tests added to cover the changes.
- [ ] Quick tests passed locally by running `./runtest.sh`.
- [ ] In-line docstrings updated.
- [ ] Documentation updated.