Convert ET model once for each task (#3616)
Fixes # .

### Description

The PT model needs to be converted to ET format before sending to a
device. There are two issues with current implementation:

- The model is converted every time a new request is received from the
device, even though the conversion is the same for all devices.
- The conversion is not done within a lock. When multiple devices ask
for the model at the same time, the multiple conversions happen
concurrently, which could break the integrity of the converted result.

This PR fixes these issues:
- The model is converted only once. Converted result is cached for later
reuse.
- The check-and-convert process is placed within a lock. 


### Types of changes
<!--- Put an `x` in all the boxes that apply, and remove the not
applicable items -->
- [x] Non-breaking change (fix or new feature that would not break
existing functionality).
- [ ] Breaking change (fix or new feature that would cause existing
functionality to change).
- [ ] New tests added to cover the changes.
- [ ] Quick tests passed locally by running `./runtest.sh`.
- [ ] In-line docstrings updated.
- [ ] Documentation updated.