Add android demo app using Android Edge SDK (#3559)
Add android demo app using the edge SDK for Android.

### Description

Add the demo app logic for android to go with the NVFlareSDK for
Android.

Routing proxy is modified to add: nvflare/edge/web/service/proto_exe.py
and allow for a config file in json for the expected payloads to receive
and send for easier testing of the mobile app

### Types of changes
<!--- Put an `x` in all the boxes that apply, and remove the not
applicable items -->
- [x] Non-breaking change (fix or new feature that would not break
existing functionality).
- [ ] Breaking change (fix or new feature that would cause existing
functionality to change).
- [ ] New tests added to cover the changes.
- [ ] Quick tests passed locally by running `./runtest.sh`.
- [ ] In-line docstrings updated.
- [ ] Documentation updated.