Make via-file decomposers configurable (#3684)
Fixes # .

### Description

Apheris has started playing with 2.7 RC and noticed slowdown in model
transfer. Though this is to be expected (since it involves creation of a
file and downloading of the file chunk by chunk), there is room for
improvement, especially if the network bandwidth is big.

This PR makes it possible to configure decomposers for optimal model
transfer performance.

The behavior of such decomposers (TensorDecomposer and
NumpyArrayDecomposer) can be configured with two parameters:

**xxx_min_file_size_for_streaming** - specifies the min model size
required for using file streaming. If the model size is less than this
size, the model will be included in the message directly, and no file
streaming is needed. If not specified, the default value is 2MB. For
systems with large bandwidth and/or memory, this parameter can take a
larger value (e.g. 1GB).

If the specified size is 0, then the object will use native
decomposition - no file is used.

**xxx_streaming_per_request_timeout** - specifies the per-request
timeout for streaming messages. If not specified, default to 10 secs.
The value of this parameter can be adjusted based on the network
performance.

Where xxx is the prefix of the specific decomposer. For
TensorDecomposer, the prefix is "tensor"; for NumpyArrayDecomposer, the
prefix is "np".

Note that by default, NumpyArrayDecomposer uses native decomposition.
TensorDecomposer does not support native decomposition.

Both parameters can be set in job config files (config_fed_server.json
and config_fed_client.json) as top-level elements.


### Types of changes
<!--- Put an `x` in all the boxes that apply, and remove the not
applicable items -->
- [x] Non-breaking change (fix or new feature that would not break
existing functionality).
- [ ] Breaking change (fix or new feature that would cause existing
functionality to change).
- [ ] New tests added to cover the changes.
- [ ] Quick tests passed locally by running `./runtest.sh`.
- [ ] In-line docstrings updated.
- [ ] Documentation updated.